fetch/metadata/generated/element-link-icon.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-link-icon.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "link" element with rel="icon"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  const params = {
    body: `
      <svg xmlns="http://www.w3.org/2000/svg" width="123" height="123">
        <rect fill="lime" width="123" height="123"/>
      </svg>
    `,
    mime: 'image/svg+xml'
  };

  /**
   * The `link` element supports a `load` event. That event would reliably
   * indicate that the browser had received the request. Multiple major
   * browsers do not implement the event, however, so in order to promote the
   * visibility of this test, a less efficient polling-based detection
   * mechanism is used.
   *
   * https://bugzilla.mozilla.org/show_bug.cgi?id=1638188
   * https://bugs.chromium.org/p/chromium/issues/detail?id=1083034
   */
  function induceRequest(t, url, attributes) {
    const link = document.createElement('link');
    link.setAttribute('rel', 'icon');
    link.setAttribute('href', url);

    for (const [ name, value ] of Object.entries(attributes)) {
      link.setAttribute(name, value);
    }

    document.head.appendChild(link);
    t.add_cleanup(() => link.remove());
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpOrigin'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
      });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-origin destination no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpSameSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
      });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-site destination no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      t,
      makeRequestURL(key, ['httpCrossSite'], params),
      {}
    );

    return retrieve(key, {poll:true})
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
      });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy cross-site destination no attributes');
  </script>
  </body>
</html>
